pyproject.toml
[tool.poetry]
name = "simple-llm-chatbot"
version = "0.1.0"
description = ""
authors = ["hungpm372 <hungpm372@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
langchain = "^0.2.6"
langchain-openai = "^0.1.10"
python-dotenv = "^1.0.1"
langchain-community = "^0.2.6"
langchain-chroma = "^0.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
